+++ app/cmd/rpc/query.go
@@ -4,8 +4,10 @@ import (
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
+	types4 "github.com/pokt-network/pocket-core/app/cmd/rpc/types"
 	sdk "github.com/pokt-network/pocket-core/types"
 	types2 "github.com/pokt-network/pocket-core/x/auth/types"
+	types3 "github.com/pokt-network/pocket-core/x/pocketcore/types"
 	abci "github.com/tendermint/tendermint/abci/types"
 	"github.com/tendermint/tendermint/libs/bytes"
 	"github.com/tendermint/tendermint/types"
@@ -22,6 +24,31 @@ import (
 
 func Version(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
 	WriteResponse(w, APIVersion, r.URL.Path, r.Host)
+}
+
+func LocalNodes(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
+	value := r.URL.Query().Get("authtoken")
+	if value != app.AuthToken.Value {
+		WriteErrorResponse(w, 401, "wrong authtoken "+value)
+		return
+	}
+	var localNodes []types4.PublicPocketNode
+	for _, node := range types3.GlobalPocketNodes {
+		if node == nil {
+			continue
+		}
+		localNodes = append(localNodes, types4.PublicPocketNode{Address: node.GetAddress().String()})
+	}
+	j, err := json.Marshal(localNodes)
+	if err != nil {
+		WriteErrorResponse(w, 400, err.Error())
+		return
+	}
+	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
+	_, err = w.Write(j)
+	if err != nil {
+		WriteErrorResponse(w, 400, err.Error())
+	}
 }
 
 type HeightParams struct {
@@ -352,6 +379,28 @@ func Account(w http.ResponseWriter, r *http.Request, ps httprouter.Params) {
 		params.Height = app.PCA.BaseApp.LastBlockHeight()
 	}
 	res, err := app.PCA.QueryAccount(params.Address, params.Height)
+	if err != nil {
+		WriteErrorResponse(w, 400, err.Error())
+		return
+	}
+	s, err := json.Marshal(res)
+	if err != nil {
+		WriteErrorResponse(w, 400, err.Error())
+		return
+	}
+	WriteJSONResponse(w, string(s), r.URL.Path, r.Host)
+}
+
+func Accounts(w http.ResponseWriter, r *http.Request, ps httprouter.Params) {
+	var params = PaginatedHeightParams{Height: 0}
+	if err := PopModel(w, r, ps, &params); err != nil {
+		WriteErrorResponse(w, 400, err.Error())
+		return
+	}
+	if params.Height == 0 {
+		params.Height = app.PCA.BaseApp.LastBlockHeight()
+	}
+	res, err := app.PCA.QueryAccounts(params.Height, params.Page, params.PerPage)
 	if err != nil {
 		WriteErrorResponse(w, 400, err.Error())
 		return