+++ doc/releases/changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 =======
+## Unreleased
+- Mempool Refinements :
+  - Tx max life in mempool is 2 blocks, after that
+  - add validation for edge case on cache equals to 0.
+- backport tendermint #6068 (terminate broadcastEvidenceRoutine when peer is stopped)
+- Add new rpc endpoint /query/accounts
+- New config.json option to enable/disable authToken generation on start (auth.json)
+
 ## RC-0.8.3
 
 - Add config.json check for Indexer and Mempool.